Tag outputs with task from which they came (#66)
To enable partitioning outputs by the task that emitted them. This comes with a breaking change in the interface to
executor, since it now emits OutputBatch instead of dataframe, but this seems low risk at the moment, and any users of
writeto will remain back-compatible. The OutputBatch structure (as opposed to tuple, for instance) was chosen to enable future metadata additions.