fixing walk order to resolve priority in multi-sink pipelines (#120)
This encourages tasks that are further along the pipeline to run as priority while still ensuring scheduling of tasks closer to the sinks. It does so by ordering tasks by their furthest distance (rather than closest) to the sink along all paths. Another change included here is to re-evaluate the same-level sort order of tasks so as to more fairly schedule ones (as opposed to submitting as many as can fit at a time of a single one).