resource request scheduling fix (#125)
For queueing tasks in the ray executor, the logic was assuming that any resource specified had >0 unit request. In fact, it is sometimes desirable to explicitly set a resource request to 0 (e.g. to override a default or be explicit), in which case we should ignore its availability in scheduling.